Split code and improve some function
