+++ code/apiParser.py
@@ -1,0 +1,27 @@
+
+from json import loads
+# convert json string to python dicitionory
+import menu
+
+def parse(dic):
+    # takes dict as param
+    # repr json data
+    # iterate over every key, and print element at that key
+    print("\n**********************")
+    for key in dic.keys():
+        print(key +" : "+str(dic[key]))
+    print("\n**********************\n")
+    input("Press Enter to continue....") 
+    menu.clear() 
+
+def listParse(list):
+    # takes list as param
+    # list contains dic
+    print("\n**********************")
+    for dic in list:
+        for key in dic.keys():
+           print(key +" : "+str(dic[key]))
+        print("\n",end="")
+    print("\n**********************\n")
+    input("Press Enter to continue....") 
+    menu.clear()
