+++ code/constant.py
@@ -1,0 +1,6 @@
+
+
+API='https://covid19-api.org/api'
+ICON_NAME="covid"
+SCHEDULE_SCRIPT="schedulescript.py"
+SCHEDULED="scheduled.txt"
