+++ code/run.py
@@ -1,0 +1,8 @@
+import menu
+
+
+def main():
+    menu.createMenu("fda","", "")
+
+
+main()
