+++ code/tracer.py
@@ -1,0 +1,107 @@
+
+from requests import get
+# make a get request
+import menu
+from json import loads
+# convert json string to python dicitionory
+import constant
+def statusCountry(country):
+    # get country name whose data is required
+    # takes API site as param
+    # returns python  dict reprsenting status of country
+    # request API using GET method
+    # convert response string to python dict
+    # if error occurs print error code and reason of error
+    site = constant.API
+    subUrl = '/status'
+    try:
+        resp=get(site+subUrl+"/"+menu.countrytoCode(country))
+    except Exception as err:
+        print("Error: " + str(err))
+        exit(1)
+    if (resp.status_code==200):
+        return(loads(resp.content))
+    else:
+        print("Error: " + str(resp.status_code) + ":" + resp.reason)
+        exit(1)
+
+def statusCountryDate(country, date):
+    # get country name and date at which status is required
+    # get date on which data is required
+    # takes API site as param
+    # returns python dict reprsenting status of country till given date
+    # request API using GET method
+    # convert response string to python dict
+    # if error occurs print error code and reason of error
+    site = constant.API
+    subUrl= '/status'
+    try:
+        resp=get(site+subUrl+"/"+menu.countrytoCode(country), params={'date':date})
+        #print(resp.url)
+    except Exception as err:
+        print("Error: " + str(err))
+        exit(1)
+    if (resp.status_code==200):
+        return(loads(resp.content))
+    else:
+        print("Error: " + str(resp.status_code) + ":" + resp.reason)
+        exit(1)
+
+def diffCountry(country):
+    # get country name
+    # takes API site as param
+    # returns python dict reprsenting difference
+    # of current case and previous case
+    # request API using GET method
+    # convert response string to python dict
+    # if error occurs print error code and reason of error
+    site = constant.API
+    try:
+        resp=get(site+"/diff/"+menu.countrytoCode(country))
+    except Exception as err:
+        print("Error: " + str(err))
+        exit(1)
+    if (resp.status_code==200):
+        return(loads(resp.content))
+    else:
+        print("Error: " + str(resp.status_code) + ":" + resp.reason)
+        exit(1)
+
+def predictionCountry(country):
+    # get country name
+    # takes API site as param
+    # returns python dict reprsenting data 
+    # of prediction of cases in next 2 weeks using ML
+    # request API using GET method
+    # convert response string to python dict
+    # if error occurs print error code and reason of error
+    site = constant.API
+    try:
+        resp=get(site+"/prediction/"+menu.countrytoCode(country))
+    except Exception as err:
+        print("Error: " + str(err))
+        exit(1)
+    if (resp.status_code==200):
+        return(loads(resp.content))
+    else:
+        print("Error: " + str(resp.status_code) + ":" + resp.reason)
+        exit(1)   
+
+def timeCases(country):
+    # get country name
+    # takes API site as param
+    # returns python dict reprsenting data of cases by country and time
+    # request API using GET method
+    # convert response string to python dict
+    # if error occurs print error code and reason of error
+    site = constant.API
+    try:
+        resp=get(site+"/timeline/"+menu.countrytoCode(country))
+    except Exception as err:
+        print("Error: " + str(err))
+        exit(1)
+    if (resp.status_code==200):
+        return(loads(resp.content))
+    else:
+        print("Error: " + str(resp.status_code) + ":" + resp.reason)
+        exit(1)
