+++ code/userInput.py
@@ -1,0 +1,13 @@
+
+
+def getChoice():
+    #Method to get user Input
+    try:
+        choice = int(input("Choice> "))
+    except ValueError:
+        print("Not a valid input")
+        return getChoice()
+    return choice
+
+def read(key):
+    return(input(key+": "))
